pyproject.toml
[project]
name = "portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.6",
    "python-docx>=1.1.2",
]
